+++ client/src/App.jsx
@@ -1,54 +1,54 @@
 import {
-	ApolloClient,
-	InMemoryCache,
-	ApolloProvider,
-	createHttpLink,
-} from "@apollo/client";
-import { setContext } from "@apollo/client/link/context";
-import { Outlet } from "react-router-dom";
+  ApolloClient,
+  InMemoryCache,
+  ApolloProvider,
+  createHttpLink,
+} from '@apollo/client'
+import { setContext } from '@apollo/client/link/context'
+import { Outlet } from 'react-router-dom'
 
-import Header from "./components/Header/Header";
+import Header from './components/Header/Header'
 // import Footer from "./components/Footer";
-import LoginButton from "./components/loginButton/loginButton"
-import LogoutButton from "./components/logoutButton/logoutButton"
+import LoginButton from './components/loginButton/loginButton'
+import LogoutButton from './components/logoutButton/logoutButton'
 
-// Construct our main GraphQL API endpoint
-const httpLink = createHttpLink({
-	uri: "/graphql",
-});
+// // Construct our main GraphQL API endpoint
+// const httpLink = createHttpLink({
+// 	uri: "/graphql",
+// });
 
-// Construct request middleware that will attach the JWT token to every request as an `authorization` header
-const authLink = setContext((_, { headers }) => {
-	// get the authentication token from local storage if it exists
-	const token = localStorage.getItem("id_token");
-	// return the headers to the context so httpLink can read them
-	return {
-		headers: {
-			...headers,
-			authorization: token ? `Bearer ${token}` : "",
-		},
-	};
-});
+// // Construct request middleware that will attach the JWT token to every request as an `authorization` header
+// const authLink = setContext((_, { headers }) => {
+// 	// get the authentication token from local storage if it exists
+// 	const token = localStorage.getItem("id_token");
+// 	// return the headers to the context so httpLink can read them
+// 	return {
+// 		headers: {
+// 			...headers,
+// 			authorization: token ? `Bearer ${token}` : "",
+// 		},
+// 	};
+// });
 
-const client = new ApolloClient({
-	// Set up our client to execute the `authLink` middleware prior to making the request to our GraphQL API
-	link: authLink.concat(httpLink),
-	cache: new InMemoryCache(),
-});
+// const client = new ApolloClient({
+// 	// Set up our client to execute the `authLink` middleware prior to making the request to our GraphQL API
+// 	link: authLink.concat(httpLink),
+// 	cache: new InMemoryCache(),
+// });
 
 function App() {
-	return (
-		<ApolloProvider client={client}>
-			<div className="mainBody">
-				<Header />
-				<div className="container">
-					<Outlet />
-				</div>
-				<LogoutButton />
-				{/* <Footer /> */}
-			</div>
-		</ApolloProvider>
-	);
+  return (
+    // <ApolloProvider client={client}>
+    <div className="mainBody">
+      <Header />
+      <div className="container">
+        <Outlet />
+      </div>
+      <LogoutButton />
+      {/* <Footer /> */}
+    </div>
+    // </ApolloProvider>
+  )
 }
 
-export default App;
+export default App
